Open British Arboricultural Record Dataset (OpenBARD)

OpenBARD are tree data used in the production of Fennell, Joseph. (2024). Handbook of UK Urban Tree Allometric Equations and Size Characteristics (Version 1.4). doi:10.13140/RG.2.2.28745.04961.

Usage

obard

Format

The data frame obard has 15357 rows and 30 variables:

record_id

Record ID

n_stems

Number of stems

stem_diameter_1_cm

Diameter in cm of first stem

stem_diameter_2_cm

Diameter in cm of second stem

stem_diameter_3_cm

Diameter in cm of third stem

stem_diameter_4_cm

Diameter in cm of fourth stem

stem_diameter_5_cm

Diameter in cm of fifth stem

height_m

Height of tree in metres

life stage

Life stage of tree

physiological_condition

Physiological condition of tree

classification

Nature of the location of the tree (urban, suburban, rural)

county

County in which tree is located

latin_name_clean

Latin (scientific) name for tree

common_name_clean

Common name for tree

generic_name_clean

Genus name for tree

genus

Genus of tree

species

Specific epithet for tree

combined_stem_diameter_cm

Combined diameter in cm of stems

is_multistem

Is the tree multistemmed?

crown_radius_n_m

Radius in metre of crown on north compass point

crown_radius_s_m

Radius in metre of crown on south compass point

crown_radius_e_m

Radius in metre of crown on east compass point

crown_radius_w_m

Radius in metre of crown on west compass point

crown_radius_se_m

Radius in metre of crown on south-east compass point

crown_radius_nw_m

Radius in metre of crown on north-west compass point

crown_radius_ne_m

Radius in metre of crown on north-east compass point

crown_radius_sw_m

Radius in metre of crown on south-west compass point

crown_radius_minimum_bounding_circle

The radius of the expected minimum bounding circle of the crown

crown_radius_maximum

The maximum radial value recorded for the tree

data_contributor_id

Dataset-unique identifier for the data contributor

Source

https://zenodo.org/records/15593688

Details

The data can be found on Zenodo: https://zenodo.org/records/15593688

Estimates for the maximum sizes of amenity (and other, non-forestry) trees are important for urban planning and tree management. This report presents the findings of an applied research study to generate current best estimates for UK tree typical sizes and allometric equations. This report contains typical mature sizes for 46 species and 29 genera of trees commonly found in the UK’s towns, cities and land outside of woodland. While the data come from sites across the UK in both urban and rural areas, the data were all collected as part of professional tree surveys and so are likely to be representative of amenity trees. Trees grown in woodland conditions are not likely to be well represented by the models presented here. This report also contains allometric model formulae for up to 23 species (depending on the relationship), allowing tree characteristics such as diameter at breast height to be predicted by other measured parameters. The formulae can be used in Excel (or other software) for prediction.

What does the document contain? (1) Summary of methods used to generate the values and equations (2) Typical and large (95th percentile) crown radius, height, diameter at breast height and root protection radius for UK amenity tree species and genera (3) Plots to show these values (4) Coefficients for single-parameter allometric equations.

Examples

# Load full data set
data(obard)

# View summary
str(obard)
#> 'data.frame':	15357 obs. of  30 variables:
#>  $ record_id                           : chr  "993206fd-a580-4649-b979-45ae23eca818" "da6f26e5-890a-41a0-96a5-469dbc1ee9b4" "85be7deb-5e04-4641-bdac-8a8c7d44c153" "998e8fb8-ebc4-4959-b060-03c5a5886b87" ...
#>  $ n_stems                             : int  1 1 2 2 1 2 2 1 1 2 ...
#>  $ stem_diameter_1_cm                  : num  55.5 65 27 24.5 40 50 35 40 55 50 ...
#>  $ stem_diameter_2_cm                  : num  NA NA 23 24 NA 45 30 NA NA 45 ...
#>  $ stem_diameter_3_cm                  : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ stem_diameter_4_cm                  : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ stem_diameter_5_cm                  : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ height_m                            : num  20 18 8 8 15 18 12 22 22 18 ...
#>  $ life_stage                          : chr  "semi_mature" "semi_mature" "mature" "mature" ...
#>  $ physiological_condition             : chr  "good" "good" "good" "good" ...
#>  $ classification                      : chr  "suburban" "suburban" "suburban" "suburban" ...
#>  $ county                              : chr  "cheshire" "cheshire" "somerset" "somerset" ...
#>  $ latin_name_clean                    : chr  "tilia europaea" "aesculus hippocastanum" "magnolia macrophylla" "magnolia macrophylla" ...
#>  $ common_name_clean                   : chr  "common lime" "horse chestnut" "bigleaf magnolia" "bigleaf magnolia" ...
#>  $ generic_name_clean                  : chr  "lime" "horse chestnut" "magnolia" "magnolia" ...
#>  $ genus                               : chr  "tilia" "aesculus" "magnolia" "magnolia" ...
#>  $ species                             : chr  "europaea" "hippocastanum" "macrophylla" "macrophylla" ...
#>  $ combined_stem_diameter_cm           : num  55.5 65 35.5 34.3 40 67.3 46.1 40 55 67.3 ...
#>  $ is_multistem                        : chr  "False" "False" "True" "True" ...
#>  $ crown_radius_n_m                    : num  8.5 8 4.5 4 7 7 5 7 7 7 ...
#>  $ crown_radius_s_m                    : num  9 9 3.5 5 7 7 5 8 8 7 ...
#>  $ crown_radius_e_m                    : num  6 6 5 5 6 6 5 7 5 5 ...
#>  $ crown_radius_w_m                    : num  6 8 4 4.5 6 6 4 2 8 8 ...
#>  $ crown_radius_se_m                   : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ crown_radius_nw_m                   : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ crown_radius_ne_m                   : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ crown_radius_sw_m                   : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ crown_radius_minimum_bounding_circle: num  8.8 8.5 4.5 4.8 7 7 5 7.5 7.5 7 ...
#>  $ crown_radius_maximum                : num  9 9 5 5 7 7 5 8 8 8 ...
#>  $ data_contributor_id                 : chr  "C1" "C1" "C1" "C1" ...